+++ tests/testthat/test-fits-examples.R
@@ -6,6 +6,6 @@ testthat::skip("This time-consuming test is only run locally before release.")
 examples = c("ar", "binomial", "demo", "group", "intercepts", "multiple", "quadratic", "variance", "varying")
 for (example in examples) {
   message("Now running example mcp_example('", example, "')")
-  fit = mcp_example(example)
+  fit = mcp_example(example, plot = FALSE)
   test_matches_simulated(fit)
 }